Move old project view into SvelteUxProjectView.svelte
