+++ frontend/viewer/src/CrdtProjectView.svelte
@@ -6,5 +6,5 @@
 </script>
 
 <ProjectLoader {projectName} let:onProjectLoaded>
-  <ProjectView {projectName} isConnected on:loaded={e => onProjectLoaded(e.detail)}></ProjectView>
+  <ProjectView {projectName} isConnected onloaded={onProjectLoaded}></ProjectView>
 </ProjectLoader>
