+++ frontend/viewer/src/TestProjectView.svelte
@@ -7,5 +7,5 @@ const projectName = InMemoryApiService.setup().projectName;
 </script>
 
 <ProjectLoader {projectName} let:onProjectLoaded>
-  <ProjectView {projectName} isConnected on:loaded={e => onProjectLoaded(e.detail)}></ProjectView>
+  <ProjectView {projectName} isConnected onloaded={onProjectLoaded}></ProjectView>
 </ProjectLoader>
